Squashed commit of develop:
commit 6f10bb4334c58c9683c6a6f513a76fa1b9961ce8
Author: Andy Fragen <andy@thefragens.com>
Date:   Sun Nov 20 09:32:00 2016 -0800

    rename function for clarity

commit 7dbd4f3c4f15363ef10f1b35989fcb8914b53c24
Author: Andy Fragen <andy@thefragens.com>
Date:   Sat Nov 19 11:23:18 2016 -0800

    split into 2 functions for single responsibility

commit 8060989616b8b3def52f925c030e70a33d933fd0
Author: Andy Fragen <andy@thefragens.com>
Date:   Sat Nov 19 10:11:39 2016 -0800

    refactor wp_update_response

    includes transient update when on plugins or themes pages

commit be4b6bfc077e22b33962fafce7b06bc9aa18d7ad
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 20:15:27 2016 -0800

    add_filter to make sure updates are shown

    after cache flush

commit 761e071280d015ce7065383ea4a31c044c16076f
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 18:36:34 2016 -0800

    move to load GHU_Upgrade only on pages GHU runs

    fix for @mlteal

commit 34963167bbb8954ceda326967450d9a53fd0ef04
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 18:28:29 2016 -0800

    type cast as array

commit 0a16a14a1729d74a168594ba5381b7f59c847d5d
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 18:27:44 2016 -0800

    simplify for dot org repos

commit 0496fbbf315605cc501c3867073f163c8a76973f
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 17:29:40 2016 -0800

    don't delete transients of error message pops up

    getting instance of class Messages runs through class Base and this avoids running the delete transients code twice.

commit 67b7f589897b0d0b2875317298a1452132ca9a74
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 16:32:49 2016 -0800

    Simplify check for extended naming

commit b55d5b970a957e18a786f4352a0242233411b227
Author: Andy Fragen <andy@thefragens.com>
Date:   Fri Nov 18 15:33:35 2016 -0800

    make plugin/theme pre_set_site_transient_update_{}

    with similar code.

commit 165fbc68d715cce31fddd277efb7d9f5cecba38c
Author: Andy Fragen <andy@thefragens.com>
Date:   Thu Nov 17 17:15:21 2016 -0800

    all theme to rollback anytime

commit 05b89ce48f97c9394b941d040af415ed1d99ab58
Author: Andy Fragen <andy@thefragens.com>
Date:   Thu Nov 17 14:51:02 2016 -0800

    Firefox can't have nice things

commit 6c64d24ea4fce068cc8d781345b3739595936817
Author: Andy Fragen <andy@thefragens.com>
Date:   Thu Nov 17 13:07:36 2016 -0800

    don't save dot org response just representation

commit 0004fbd58e1cb3d254ce656501d9b554831a3e2d
Author: Andy Fragen <andy@thefragens.com>
Date:   Thu Nov 17 08:36:15 2016 -0800

    readme update

commit 180f09ee172f590119a03967dc11cd775ae4a541
Author: Andy Fragen <andy@thefragens.com>
Date:   Thu Nov 17 08:28:25 2016 -0800

    eliminate loop

commit a17f839c8eb5d0015f1abe902fd66fc9ab0c726b
Author: Andy Fragen <andy@thefragens.com>
Date:   Wed Nov 16 21:04:27 2016 -0800

    comment

commit 93921c9de776619465ccf3a4e0a6d30c63c01dbc
Author: Andy Fragen <andy@thefragens.com>
Date:   Wed Nov 16 21:01:31 2016 -0800

    oops, make sure dot org only plugins return info

commit 55ec4c744f76dc5df82d04ccd38c5e693f1382c9
Author: Andy Fragen <andy@thefragens.com>
Date:   Wed Nov 16 20:54:11 2016 -0800

    major refactor for checking wp.org for plugin data

    this is what happens when I have to wait for an operating room for hours. :P

commit 24bf89cae6d574099b5f148420dde92f17462b79
Author: Andy Fragen <andy@thefragens.com>
Date:   Wed Nov 16 20:49:27 2016 -0800

    update readme for Slack info

commit be8b40f87390b676cdc9fb74a559b1b36a0d5e09
Author: Andy Fragen <andy@thefragens.com>
Date:   Wed Nov 16 10:29:13 2016 -0800

    switch to Heroku app, https://github.com/rauchg/slackin

commit 5affc68351ddcf8575e75fb6bbcd3f696b475ccb
Author: Andy Fragen <andy@thefragens.com>
Date:   Tue Nov 15 18:26:40 2016 -0800

    yeah, these weren't needed for dot org fix

commit 05aede8de2850d1d380aa6381d6ce26d680f9910
Author: Andy Fragen <andy@thefragens.com>
Date:   Tue Nov 15 17:58:10 2016 -0800

    fix for mismatched versions on dot org plugins

    https://github.com/afragen/github-updater/issues/496

commit e42c194afea3851aeb5aff5e922c9bb3f69134ec
Author: Andy Fragen <andy@thefragens.com>
Date:   Tue Nov 15 12:24:20 2016 -0800

    encourage curiosity

commit 1153a534bc0489f8945388c37d12960b274e6b24
Author: Andy Fragen <andy@thefragens.com>
Date:   Tue Nov 15 10:42:18 2016 -0800

    update for Slack invites via stacktodo bot

commit a439313de6feddb86ebc926e273db126d97c40d1
Author: Andy Fragen <andy@thefragens.com>
Date:   Sun Nov 13 19:15:13 2016 -0800

    always check tags

    otherwise download link by revert to branch

commit 0949e720c88a85f24eec88e21256c88471473c4e
Author: Andy Fragen <andy@thefragens.com>
Date:   Sun Nov 13 18:45:42 2016 -0800

    fix PHP notice

+++ src/GitHub_Updater/Base.php
@@ -137,12 +137,9 @@ class Base {
 	 * Loads options to private static variable.
 	 */
 	public function __construct() {
-		if ( isset( $_POST['ghu_refresh_cache'] ) ) {
+		if ( isset( $_POST['ghu_refresh_cache'] ) && ! ( $this instanceof Messages ) ) {
 			$this->delete_all_transients();
 		}
-
-		// Run GitHub Updater upgrade functions.
-		new GHU_Upgrade();
 
 		$this->load_hooks();
 	}
@@ -240,6 +237,12 @@ class Base {
 
 		if ( in_array( $pagenow, array_unique( $admin_pages ) ) ) {
 			$force_meta_update = true;
+
+			// Run GitHub Updater upgrade functions.
+			new GHU_Upgrade();
+
+			// Ensure transient updated on plugins.php and themes.php pages.
+			add_action( 'admin_init', array( &$this, 'admin_pages_update_transient' ) );
 		}
 
 		if ( isset( $_POST['ghu_refresh_cache'] ) ) {
@@ -893,9 +896,6 @@ class Base {
 
 		$wpdb->query( $wpdb->prepare( $delete_string, array( '%_ghu-%' ) ) );
 
-		set_site_transient( 'update_plugins', null );
-		set_site_transient( 'update_themes', null );
-
 		return true;
 	}
 
@@ -1317,6 +1317,29 @@ class Base {
 			$updates_transient->response[ $slug ] = (array) $rollback;
 		}
 		set_site_transient( $transient, $updates_transient );
+	}
+
+	/**
+	 * Ensure update transient is update to date on admin pages.
+	 */
+	public function admin_pages_update_transient() {
+		global $pagenow;
+
+		$admin_pages   = array( 'plugins.php', 'themes.php' );
+		$is_admin_page = in_array( $pagenow, $admin_pages ) ? true : false;
+		$capability    = 'update_' . rtrim( $pagenow, '.php' );
+
+		if ( current_user_can( $capability ) && $is_admin_page ) {
+			$current = get_site_transient( $capability );
+			if ( 'plugins.php' === $pagenow ) {
+				$current = Plugin::instance()->pre_set_site_transient_update_plugins( $current );
+			}
+			if ( 'themes.php' === $pagenow ) {
+				$current = Theme::instance()->pre_set_site_transient_update_themes( $current );
+			}
+			set_site_transient( $capability, $current );
+		}
+		remove_filter( 'admin_init', array( &$this, 'admin_pages_update_transient' ) );
 	}
 
 	/**
